+++ TLRPC$TL_chatParticipantsForbidden_old.java
@@ -5,12 +5,12 @@ public class TLRPC$TL_chatParticipantsForbidden_old extends TLRPC$TL_chatPartici
 
     @Override // org.telegram.tgnet.TLObject, org.telegram.tgnet.TLRPC$TL_chatParticipantsForbidden
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
-        this.chat_id = abstractSerializedData.readInt32(z);
+        this.chat_id = (long) abstractSerializedData.readInt32(z);
     }
 
     @Override // org.telegram.tgnet.TLObject, org.telegram.tgnet.TLRPC$TL_chatParticipantsForbidden
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.chat_id);
+        abstractSerializedData.writeInt32((int) this.chat_id);
     }
 }
