+++ TLRPC$TL_decryptedMessageMediaContact.java
@@ -8,7 +8,7 @@ public class TLRPC$TL_decryptedMessageMediaContact extends TLRPC$DecryptedMessag
         this.phone_number = abstractSerializedData.readString(z);
         this.first_name = abstractSerializedData.readString(z);
         this.last_name = abstractSerializedData.readString(z);
-        this.user_id = abstractSerializedData.readInt32(z);
+        this.user_id = (long) abstractSerializedData.readInt32(z);
     }
 
     @Override // org.telegram.tgnet.TLObject
@@ -17,6 +17,6 @@ public class TLRPC$TL_decryptedMessageMediaContact extends TLRPC$DecryptedMessag
         abstractSerializedData.writeString(this.phone_number);
         abstractSerializedData.writeString(this.first_name);
         abstractSerializedData.writeString(this.last_name);
-        abstractSerializedData.writeInt32(this.user_id);
+        abstractSerializedData.writeInt32((int) this.user_id);
     }
 }
